# docs/swagger.yaml
basePath: /
definitions:
  kasir-api_internal_dto.CategoryRequest:
    properties:
      description:
        maxLength: 255
        type: string
      name:
        minLength: 1
        type: string
    required:
    - name
    type: object
  kasir-api_internal_dto.ProductRequest:
    properties:
      name:
        minLength: 1
        type: string
      price:
        type: integer
      stock:
        type: integer
    required:
    - name
    - price
    - stock
    type: object
host: https://kasir-api-production-1c80.up.railway.app/
info:
  contact: {}
  title: Kasir API
  version: "1.0"
paths:
  /api/categories:
    get:
      consumes:
      - application/json
      description: Mengambil semua data kategori
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Membuat kategori baru
      parameters:
      - description: Category Data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/kasir-api_internal_dto.CategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new category
      tags:
      - categories
  /api/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus kategori berdasarkan ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete category
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Mengambil kategori berdasarkan ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update kategori berdasarkan ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category Data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/kasir-api_internal_dto.CategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update category
      tags:
      - categories
  /api/products:
    get:
      consumes:
      - application/json
      description: Mengambil semua data produk
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get all products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Membuat produk baru
      parameters:
      - description: Product Data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/kasir-api_internal_dto.ProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new product
      tags:
      - products
  /api/products/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus produk berdasarkan ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete product
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Mengambil produk berdasarkan ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get product by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update produk berdasarkan ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product Data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/kasir-api_internal_dto.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update product
      tags:
      - products
schemes:
- http
- https
swagger: "2.0"